Eu.Fittest.Logging.Compression.RawlogParser

Description

This module provides the parser to parse FITTEST raw logs.

Documentation

data RawLogEntry timestamp attrib paragraph

A representation of entries in a raw-log file. We will make it more generic than needed so that we can reuse it later.

Constructors

Section (Maybe timestamp) attrib [RawLogEntry timestamp attrib paragraph]
Section

Par paragraph
Paragraph

parseLog :: Parser [RawLogEntry UTCTimeStamp1970 String [String]]

isParagraph :: RawLogEntry t t1 t2 -> Bool

isSection :: RawLogEntry t t1 t2 -> Bool

strParseRawLog :: Text -> [RawLogEntry UTCTimeStamp1970 String [String]]

To parse a string containing raw log entries.

printStatistics :: String -> IO ()

type UTCTimeStamp1970 = (Int, Integer)

Representation of UTC timestamp as two integer (o,t) where o is the location offset in minutes, and t is time in terms of elapses millisecond after midnight of 1st jan 1970.